Added search drop down component to search component. Moved tests to proper files.
